feat(inode): implement inode collection
To be compatible with iget5_locked we need at

least inode collection trait. However, iget5_locked

allocs the inode first and fills it instead of using RAII styles.

We have to use MaybeUninit to init separately.

On filesystem we need factor the collection out and use a separate function

to do the inode reading stuff because the borrow checker cannot check it.

Signed-off-by: Yiyang Wu <toolmanp@tlmp.cc>